Moved code into module terrachecker.
