+++ setup.py
@@ -1,0 +1,14 @@
+#!/usr/bin/env python
+
+from distutils.core import setup
+
+
+setup(
+    name='terrachecker',
+    version='0.1',
+    description='Terraform checker.',
+    author='Mark Winterbottom',
+    author_email='mark.winterbottom@jisc.ac.uk',
+    url='https://github.com/JiscRDSS/rdss-terraform-standards-checker',
+    packages=['terrachecker'],
+)
